OV-19. Make coap server a stand-alone component. Adapt JRC and OpenBenchmark agent.
