+++ bin/logging.conf
@@ -41,7 +41,7 @@ formatter=benchmark
 #============================ loggers =========================================
 
 [loggers]
-keys=root,eventBusMonitor,openTun,openTunWindows,openTunLinux,eventBusClient,lbrClient,moteConnector,moteProbe,moteProbeUtils,moteState,openLbr,openBenchmarkAgent,OpenParser,Parser,OpenHdlc,ParserData,ParserPrintf,ParserInfoErrorCritical,ParserStatus,RPL,SourceRoute,JRC,networkEventLogger,openVisualizerApp,openVisualizerGui,openVisualizerCli,openVisualizerWeb,OVtracer
+keys=root,eventBusMonitor,openTun,openTunWindows,openTunLinux,eventBusClient,lbrClient,moteConnector,moteProbe,moteProbeUtils,moteState,openLbr,openBenchmarkAgent,OpenParser,Parser,OpenHdlc,ParserData,ParserPrintf,ParserInfoErrorCritical,ParserStatus,RPL,SourceRoute,JRC,networkEventLogger,openVisualizerApp,openVisualizerGui,openVisualizerCli,openVisualizerWeb,OVtracer,coapServer
 
 [logger_root]
 level=ERROR
@@ -220,4 +220,10 @@ level=DEBUG
 handlers=std
 propagate=0
 qualname=OVtracer
+
+[logger_coapServer]
+level=INFO
+handlers=std
+propagate=0
+qualname=coapServer
